requirements.txt
Flask>=2.0
openai>=0.27.0
flask-cors>=3.0.10
python-dotenv>=1.0.0
